api: fix result ordering
Previously, order_by sorted only JSON objects, *after* query offsets and
limits, which was incorrect behavior. DB query set was in undefined
order and the sorting affected only a subset of the query set. Now,
order_by sorts the whole query set and hence the sorting is handled in
the DB layer.