api: ensure agenda_item.origin_last_modified_time is not null
origin_last_modified_time is not always availabe and might be null in
the database. Applications however seem to expect that it's always
available let's just return the last_modified_time as
origin_last_modified_time in case its null.

last_modified_time = the time of the last modification of the db row
origin_last_modified_time = the time of the last modification of the original source